packages/backend/pyproject.toml
[project]
name = "tagify-backend"
version = "0.1.0"
description = "Tagify backend API server (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]",
  "uvicorn",
  "pymongo",
  "pillow",
  "python-dotenv"
]

#[tool.uv]
# uv configuration can be added here if needed.

